reworked main.py and main_methods.py; refactored large code blocks into meaningful smaller methods
