+++ main_methods.py
@@ -4,16 +4,13 @@ import os
 import pathlib
 import pickle
 import shutil
-import uuid
-from collections import OrderedDict, defaultdict, namedtuple
+from collections import defaultdict
 from inspect import getfullargspec
 from pydoc import locate
-from time import sleep
 from typing import Union, Iterable, Optional, cast
 
 import flask
 import networkx as nx
-import numpy as np
 import requests
 import yaml
 from flask import request, jsonify, render_template_string, Response
@@ -32,131 +29,27 @@ from main_utils import (
     HTTPResponses,
     StepsName,
     pipeline_query_params,
-    steps_relation_dict,
-    add_status_to_running_process,
     PipelineLanguage,
     get_bool_expression,
     StoppableThread,
     string_conformity,
     BaseUtil,
-    transform_document_addition_results,
     PersistentObjects,
 )
 from src.rag.TextSplitters import PreprocessedSpacyTextSplitter
 from src.rag.embedding_stores.AbstractEmbeddingStore import ChunkEmbeddingStore
 from src import data_functions, cluster_functions, embedding_functions
-from src.graph_functions import GraphIncorp
-from src.util_functions import DocumentStore, EmbeddingStore, save_pickle
 
 
-pipeline_json_config = namedtuple(
-    "pipeline_json_config",
-    [
-        "name",
-        "language",
-        "document_server",
-        "vectorstore_server",
-        "data",
-        "embedding",
-        "clustering",
-        "graph",
-    ],
+from src.api.request_parsing import (
+    document_adding_json,
+    get_doc_ids,
+    parse_document_adding_json,
+    parse_pipeline_config_json,
+    parse_rag_config_json,
+    pipeline_json_config,
+    rag_config_json,
 )
-
-document_adding_json = namedtuple(
-    "document_adding_json",
-    ["language", "documents", "document_server", "vectorstore_server"],
-)
-
-
-rag_config_json = namedtuple(
-    "rag_config_json",
-    ["chatter", "api_key", "language", "prompt_template", "vectorstore_server"],
-)
-
-
-def parse_pipeline_config_json(response_json) -> pipeline_json_config:
-    config = Munch.fromDict(response_json)
-    try:
-        return pipeline_json_config(
-            string_conformity(config.get("name", None)),
-            config.get("language", None),
-            config.get("document_server", None),
-            config.get(
-                "vectorstore_server",
-                config.get(
-                    "vector_store_server", config.get("vector-store_server", None)
-                ),
-            ),
-            config.config.get("data", Munch()),
-            config.config.get("embedding", Munch()),
-            config.config.get("clustering", Munch()),
-            config.config.get("graph", Munch()),
-        )
-    except AttributeError as e:
-        logging.error(
-            f"Json body/configuration seems to be malformed: no 'config' entry was provided.\n{e}"
-        )
-        return pipeline_json_config(
-            string_conformity(config.get("name", None)),
-            config.get("language", None),
-            config.get("document_server", None),
-            config.get("vectorstore_server", None),
-            Munch(),
-            Munch(),
-            Munch(),
-            Munch(),
-        )
-
-
-def parse_document_adding_json(response_json) -> Optional[document_adding_json]:
-    try:
-        config = Munch.fromDict(response_json)
-        # _id_key = list(set(config.keys()).intersection(["id", "_id"]))
-        # _id = _id_key[0] if _id_key else "none"
-        return document_adding_json(
-            config.get("language", None),
-            config.get("documents", []),
-            config.get("document_server", None),
-            config.get("vectorstore_server", None),
-        )
-    except Exception as e:
-        logging.error(f"Content json parsing error: '{e}'")
-        return None
-
-
-def parse_rag_config_json(response_json) -> Optional[rag_config_json]:
-    try:
-        config = Munch.fromDict(response_json)
-        _chatter = config.get("chatter", None)
-        _api_key = config.get("api_key", "")
-        _lang = config.get("language", "en")
-        _prompt = config.get("prompt_template", {})
-        _vs = config.get(
-            "vectorstore_server",
-            config.get("vector_store_server", config.get("vector-store_server", None)),
-        )
-        return rag_config_json(
-            {} if (_chatter is None or not isinstance(_chatter, dict)) else _chatter,
-            _api_key,
-            _lang,
-            None if (_prompt is not None and len(_prompt) == 0) else _prompt,
-            _vs,
-        )
-    except Exception as e:
-        logging.error(f"Content json parsing error: '{e}'")
-        return None
-
-
-def get_doc_ids(response_json: dict):
-    try:
-        if intersection := {"doc_ids", "doc_id", "ids", "id"}.intersection(
-            k for k in response_json.keys()
-        ):
-            return response_json.get(list(intersection)[0])
-    except Exception as e:
-        logging.warning(f"Couldn't get document ids from request json: '{e}'")
-    return []
 
 
 def get_pipeline_query_params(
@@ -350,184 +243,13 @@ def get_documents_from_es_server(
                 yield check_es_source_for_id(document, other_id)
 
 
-def populate_running_processes(
-    app: flask.Flask, path: Union[str, pathlib.Path], running_processes: dict
-):
-    for process in get_all_processes(path):
-        _finished = [
-            _finished_step.get("name") for _finished_step in process.get("status", [])
-        ]
-        _process_name = process.get("name", None)
-        if _process_name is None:
-            app.logger.warning(
-                f"Skipping process entry with no name and '{_finished}' steps."
-            )
-            continue
-
-        for _step, _rank in steps_relation_dict.items():
-            if _step not in _finished:
-                add_status_to_running_process(
-                    _process_name, _step, ProcessStatus.NOT_PRESENT, running_processes
-                )
-            else:
-                add_status_to_running_process(
-                    _process_name, _step, ProcessStatus.FINISHED, running_processes
-                )
-    return running_processes
-
-
-def get_all_processes(path: Union[str, pathlib.Path]):
-    if isinstance(path, str):
-        path = pathlib.Path(path)
-    _process_detailed = list()
-    for _proc in path.glob("*"):
-        if _proc.is_dir() and not _proc.stem.startswith("."):
-            _proc_name = _proc.stem.lower()
-            _steps_list = list()
-            for _pickle in pathlib.Path(path / _proc_name).glob("*.pickle"):
-                _pickle_stem = _pickle.stem.lower()
-                _step = _pickle_stem.removeprefix(f"{_proc_name}_")
-                if steps_relation_dict.get(_step, False):
-                    _steps_list.append(
-                        {
-                            "rank": steps_relation_dict.get(_step),
-                            "name": _step,
-                            "status": ProcessStatus.FINISHED,
-                        }
-                    )
-            _ord_dict = OrderedDict()
-            _ord_dict["name"] = _proc_name
-            _ord_dict["status"] = sorted(_steps_list, key=lambda x: x.get("rank", -1))
-            _process_detailed.append(_ord_dict)
-    return _process_detailed
-
-
-def start_processes(
-    app: flask.Flask,
-    processes: tuple,
-    process_name: str,
-    process_tracker: dict[str, dict],
-    thread_store: dict[str, StoppableThread],
-    active_process_objs: dict[str, dict],
-    last_step: str,
-):
-    _name_marker = {
-        StepsName.DATA: "**data**, embedding, clustering, graph, integration",
-        StepsName.EMBEDDING: "data, **embedding**, clustering, graph, integration",
-        StepsName.CLUSTERING: "data, embedding, **clustering**, graph, integration",
-        StepsName.GRAPH: "data, embedding, clustering, **graph**, integration",
-        StepsName.INTEGRATION: "data, embedding, clustering, graph, **integration**",
-    }
-    for process_obj, _fact, _name in processes:
-        process_obj: BaseUtil
-        this_thread: Optional[StoppableThread] = thread_store.get(process_name, None)
-        if this_thread is not None and this_thread.set_to_stop:
-            add_status_to_running_process(
-                process_name=process_name,
-                step_name=_name,
-                step_status=ProcessStatus.NOT_PRESENT,
-                running_processes=process_tracker,
-            )
-            continue
-        log_warning = f"Something went wrong with one of the previous steps: {_name_marker[_name]}."
-        if any(
-            [
-                True
-                for d in process_tracker.get(process_name, {}).get("status", [])
-                if (d.get("name") == _name and d.get("status") == ProcessStatus.ABORTED)
-            ]
-        ):
-            app.logger.warning(log_warning + f"\n So this one was aborted: '{_name}'.")
-            continue
-        try:
-            if this_thread.set_to_stop:
-                app.logger.info(
-                    f"Thread for '{process_name}' was stopped before this step '{_name}'. So subsequent steps will not be started."
-                )
-                return
-            process_obj.start_process(
-                cache_name=process_name,
-                process_factory=_fact,
-                process_tracker=process_tracker,
-                active_process_objs=active_process_objs,
-                thread=this_thread,
-            )
-            if _name == last_step:
-                app.logger.info(f"Pipeline finished with last step: '{_name}'.")
-        except FileNotFoundError as e:
-            app.logger.warning(log_warning + f"\nThere is a pickle file missing: {e}")
-
-
-def start_thread(
-    app: flask.Flask,
-    process_name: str,
-    pipeline_thread: StoppableThread,
-    threading_store: Optional[dict[str, StoppableThread]],
-):
-    app.logger.info(f"Starting thread for '{process_name}'.")
-    pipeline_thread.start()
-    # if threading_store is not None: threading_store[process_name] = pipeline_thread
-    sleep(1.5)
-    return True
-
-
-def stop_thread(
-    app: flask.Flask,
-    process_name: str,
-    threading_store: dict[str, StoppableThread],
-    process_tracker: dict[str, dict],
-    hard_stop: bool = False,
-):
-    # ToDo: delete config for aborted steps
-    # ToDo: maybe allow for hard stop? -> terminate the thread even if a step is not yet finished
-    app.logger.info(f"Trying to stop thread for '{process_name}'.")
-    if hard_stop:
-        app.logger.warning(f"Hard stopping a thread is not implemented.")
-        hard_stop = False
-    _thread = threading_store.get(process_name, None)
-    _process = process_tracker.get(process_name, None)
-
-    if _thread is None or _process is None:
-        _msg = f"No thread/process for '{process_name}' was found."
-        logging.error(_msg)
-        return Response(_msg, status=int(HTTPResponses.INTERNAL_SERVER_ERROR))
-
-    _current_step = next(
-        (
-            step
-            for step in sorted(
-                _process.get("status", {}), key=lambda p: p.get("rank", 99)
-            )
-            if step.get("status", None) == ProcessStatus.RUNNING
-        ),
-        None,
-    )
-    if _current_step is None:
-        _msg = f"Couldn't find a running step in the pipeline '{process_name}'."
-        logging.error(_msg)
-        return Response(_msg, status=int(HTTPResponses.INTERNAL_SERVER_ERROR))
-    _thread.stop(hard_stop=hard_stop)
-
-    try:
-        for _step in sorted(
-            _process.get("status", {}), key=lambda p: p.get("rank", 99)
-        ):
-            if _step.get("rank") <= _current_step.get("rank"):
-                if _step.get("name") == _current_step.get("name"):
-                    process_tracker[process_name]["status"][_step.get("rank") - 1][
-                        "status"
-                    ] = ProcessStatus.STOPPED
-                continue
-            process_tracker[process_name]["status"][_step.get("rank") - 1][
-                "status"
-            ] = ProcessStatus.ABORTED
-    except Exception:
-        pass
-
-    app.logger.info(
-        f"Thread for '{process_name}' will be stopped after the present step ('{_current_step.get('name', None)}') is completed."
-    )
-    return Response("Process will be stopped.", status=int(HTTPResponses.ACCEPTED))
+from src.pipeline.processes import (
+    get_all_processes,
+    populate_running_processes,
+    start_processes,
+    start_thread,
+    stop_thread,
+)
 
 
 def read_config(
@@ -815,200 +537,7 @@ def get_omit_pipeline_steps(steps: object) -> list[str]:
     return []
 
 
-def add_documents_to_concept_graphs(
-        #ToDo?: ``store_permanently`` only changes whether new phrases will be stored in graphs!
-        #   --> regardless of the former argument:
-        #       - docs won't be stored in the processed_data
-        #       - docs (their phrase embeddings) will be stored in the vector store
-    content_json: document_adding_json,
-    data_processing: Optional[
-        data_functions.DataProcessingFactory.DataProcessing
-    ] = None,
-    embedding_processing: Optional[
-        embedding_functions.SentenceEmbeddingsFactory.SentenceEmbeddings
-    ] = None,
-    graph_processing: Optional[list[nx.Graph]] = None,
-    storage_path: Optional[Union[str, pathlib.Path]] = None,
-    process_name: str = "default",
-    store_permanently: bool = True,
-    document_store_cls: str = "src.marqo_external_utils.MarqoDocumentStore",
-    embedding_store_cls: str = "src.marqo_external_utils.MarqoEmbeddingStore",
-    document_cls: str = "src.marqo_external_utils.MarqoDocument",
-):
-    try:
-        document_store = locate(document_store_cls)
-        embedding_store = locate(embedding_store_cls)
-        document = locate(document_cls)
-
-        if content_json.documents is None:
-            return {"error": "No content provided."}, HTTPResponses.BAD_REQUEST
-
-        ###
-        try:
-            data_processing = (
-                FactoryLoader.load_data(str(storage_path.resolve()), process_name)
-                if data_processing is None
-                else data_processing
-            )
-            embedding_processing = (
-                FactoryLoader.load_embedding(
-                    str(storage_path.resolve()),
-                    process_name,
-                    data_processing,
-                    (
-                        None
-                        if content_json.vectorstore_server is None
-                        else content_json.vectorstore_server
-                    ),
-                )
-                if embedding_processing is None
-                else embedding_processing
-            )
-        except FileNotFoundError as e:
-            _missing = "data" if data_processing is None else "embedding"
-            return {
-                "error": f"The serialized object for '{_missing}' doesn't seem to be present. Please finish the complete pipeline for the process '{process_name}' first."
-            }, HTTPResponses.NOT_FOUND
-        has_graph = True
-        try:
-            graph_processing = (
-                FactoryLoader.load_graph(str(storage_path.resolve()), process_name)
-                if graph_processing is None
-                else graph_processing
-            )
-        except FileNotFoundError as e:
-            has_graph = False
-            logging.warning(
-                f"The serialized object for 'graph' doesn't seem to be present. Storing the document into the vector store will still be performed."
-            )
-        if (
-            content_json.vectorstore_server is None
-            and embedding_processing.source is None
-        ):
-            return {
-                "error": "Only adding documents with a vectorstore server setup is supported; no vectorstore configured."
-            }, HTTPResponses.NOT_IMPLEMENTED
-        if embedding_processing.source is None:
-            embedding_processing.source = content_json.vectorstore_server
-        if len(content_json.documents) > 0 and isinstance(
-            content_json.documents[0], dict
-        ):
-            # ToDo
-            pass
-        else:
-            return {
-                "error": "Right now only processing of documents as json is supported."
-            }, HTTPResponses.NOT_IMPLEMENTED
-        ###
-
-        _source = embedding_processing.source
-        _client_key = (
-            list(
-                set(_source.keys()).intersection(
-                    ["client_url", "url", "client", "clienturl"]
-                )
-            )
-            if isinstance(_source, dict)
-            else "none"
-        )
-        _client_key = _client_key[0] if len(_client_key) > 0 else None
-        _index_key = (
-            list(set(_source.keys()).intersection(["index_name", "index", "indexname"]))
-            if isinstance(_source, dict)
-            else "none"
-        )
-        _index_key = _index_key[0] if len(_index_key) > 0 else None
-
-        doc_content = lambda x: x if isinstance(x, dict) else {"content": x}
-        _chunk_result = data_processing.process_external_docs(
-            content=[
-                {
-                    "id": doc_content(doc).get("id", str(uuid.uuid4())),
-                    "name": doc_content(doc).get("name", None),
-                    "content": doc_content(doc).get("content", ""),
-                    "label": doc_content(doc).get("label", None),
-                }
-                for doc in content_json.documents
-            ]
-        )
-        text_list = []
-        idx_dict = defaultdict(list)
-        for idx, _chunk in enumerate(
-            # sorted(
-            (
-                (
-                    _result["text"],
-                    set(_doc["id"] for _doc in _result["doc"]),
-                )
-                for _result in _chunk_result
-            )  # ,
-            # key=lambda _result: _result[0],
-            # )
-        ):
-            for _doc in _chunk[1]:
-                idx_dict[_doc].append(idx)
-            text_list.append(_chunk[0])
-        _embedding_result = embedding_processing.encode_external(content=text_list)
-
-        embedding_store_impl: EmbeddingStore = embedding_store(
-            client_url=_source.get(_client_key, "http://localhost:8882"),
-            index_name=_source.get(_index_key, "default"),
-            create_index=False,
-            vector_dim=embedding_processing.embedding_dim,
-        )
-        doc_store_impl: DocumentStore = document_store(
-            embedding_store=embedding_store_impl
-        )
-        added_embeddings = doc_store_impl.add_documents(
-            [
-                (
-                    document(
-                        phrases=np.take(text_list, idx, 0),
-                        embeddings=np.take(_embedding_result.astype("float64"), idx, 0),
-                        doc_id=_id,
-                    ),
-                    {
-                        "offsets": [
-                            x.get("offsets", [])
-                            for _dict in np.take(_chunk_result, idx, 0)
-                            for x in _dict.get("doc", [])
-                            if x.get("id", "") == _id
-                        ],
-                        "text": [
-                            _dict.get("text", "")
-                            for _dict in np.take(_chunk_result, idx, 0)
-                        ],
-                    },
-                )
-                for _id, idx in idx_dict.items()
-            ],
-            as_tuple=True,
-        )
-        if store_permanently:
-            graph_storage_path = pathlib.Path(
-                storage_path / f"{process_name}_{StepsName.GRAPH}"
-            ).resolve()
-            save_pickle(
-                GraphIncorp.with_graphs(graph_processing)
-                .incorporate_phrases(
-                    transform_document_addition_results(
-                        (
-                            (
-                                k,
-                                v.get("with_graph"),
-                            )
-                            for k, v in added_embeddings.items()
-                        )
-                    ).items()
-                )
-                .graphs,
-                graph_storage_path,
-            )
-    except Exception as e:
-        return {
-            "error": str(e) + "\n--- please consult the logs!"
-        }, HTTPResponses.INTERNAL_SERVER_ERROR
-    return added_embeddings, HTTPResponses.OK
+from src.pipeline.document_addition import add_documents_to_concept_graphs
 
 
 def delete_pipeline(
